+++ packages/core/src/lib/BannerStrip/types.ts
@@ -1,15 +1,26 @@
+import React from 'react';
 import { ButtonProps } from '../Button';
 
-export interface BannerStripProps {
+export interface IBannerStripProps {
     /**
-     * do you want to display a button
+     * @deprecated do you want to display a button
      */
     buttonDisplayed?: boolean;
 
     /**
-     * set Button props
+     * @deprecated set Button props
      */
     buttonConfig?: ButtonProps;
+
+    /**
+     * set id to differentiate between different banner on your page
+     */
+    id: string;
+
+    /**
+     * set custom border radius
+     */
+    borderRadius: string;
 
     /**
      * set custom height
@@ -17,12 +28,49 @@ export interface BannerStripProps {
     height?: string;
 
     /**
-     * please add the text you want to show in the banner
+     * set custom width
      */
-    text: string;
+    width?: string;
+
+    /**
+     * please add the text or any custom component you want to show in the banner
+     */
+    text: React.ReactNode;
 
     /**
      * you can set the type of banner which changes its color
      */
-    type?: 'error' | 'standard' | 'success' | 'warning';
+    type?: TBannerStripTypes;
+
+    /**
+     * set custom color only when type = custom
+     */
+    bgColor?: string;
+
+    /**
+     * to display close button
+     */
+    isCloseBtnVisible?: boolean;
+
+    /**
+     * close button callback function
+     */
+    onCloseBtnClick?: () => void;
+
+    /**
+     * to set position of the banner strip
+     */
+    position?: 'fixed' | 'absolute' | 'relative';
+
+    /**
+     * specify the number of days you want to hide the component after close button is clicked (Uses local storage)
+     */
+    noOfDaysToHide?: number | null;
 }
+
+export type TBannerStripTypes =
+    | 'error'
+    | 'standard'
+    | 'success'
+    | 'warning'
+    | 'custom';