+++ packages/core/src/lib/CryptoLogos/CryptoLogos.stories.tsx
@@ -71,3 +71,8 @@ Ronin.args = {
     chain: 'ronin',
     size: '48px',
 };
+export const Optimism = Template.bind({});
+Optimism.args = {
+    chain: 'optimism',
+    size: '48px',
+};