+++ packages/core/src/lib/Illustrations/Illustration.stories.tsx
@@ -18,7 +18,7 @@ const BackgroundColoredTemplate: ComponentStory<typeof Illustration> = (
         style={{
             width: '90vw',
             height: '90vh',
-            backgroundColor: `${color.blue30}`,
+            backgroundColor: `${color.navy80}`,
         }}
     >
         <Illustration {...args} />
@@ -68,6 +68,10 @@ Cronos.args = {
 export const Ronin = BackgroundColoredTemplate.bind({});
 Ronin.args = {
     logo: 'ronin',
+};
+export const Optimism = BackgroundColoredTemplate.bind({});
+Optimism.args = {
+    logo: 'optimism',
 };
 
 export const Crypto = BackgroundColoredTemplate.bind({});